spin_plugins

Module manager

Structs

PluginManager

Provides accesses to functionality to inspect and manage the installation of plugins.

Enums

InstallAction

The action required to install a plugin to the desired version.

ManifestLocation

Location of manifest of the plugin to be installed.

PluginManagerUpdateGuard

PluginManagerUpdateLock

Functions

get_package

Gets the appropriate package for the running OS and Arch if exists