spin_plugins

Module manifest

Structs

PluginManifest

Expected schema of a plugin manifest. Should match the latest Spin plugin manifest JSON schema: https://github.com/spinframework/spin-plugins/tree/main/json-schema

PluginPackage

Describes compatibility and location of a plugin source.

Functions

warn_unsupported_version

Checks whether the plugin supports the currently running version of Spin and prints a warning if not (or if uncertain).